backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
pymongo==4.4.0
PyJWT==2.8.0
bcrypt==4.0.1
python-dotenv==1.0.0
scikit-learn==1.3.0
numpy==1.24.0
